+++ frontend/src/main.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import ReactDOM from "react-dom/client";
 import App from "./App";
 import { ErrorBoundary } from "react-error-boundary";
-import ErrorFallback from "./components/ErrorFallback";
+import ErrorFallback from "./components/UI/ErrorFallback";
 import "./index.css";
 
 ReactDOM.createRoot(document.getElementById("root")).render(
